move influxdump command code to bin module
